backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
motor>=3.2.0
pydantic>=2.0.0
python-multipart>=0.0.6
httpx>=0.24.1
dnspython>=2.3.0
python-telegram-bot>=20.0
